Forgot some definitions in 13ff659701f84d067f14060d1dffb62ee57c12de

+++ src/HLW8012.cpp
@@ -155,14 +155,14 @@ void HLW8012::expectedCurrent(double value) {
     if (_current > 0) _current_multiplier *= (value / _current);
 }
 
-void HLW8012::expectedVoltage(unsigned int value) {
+void HLW8012::expectedVoltage(double value) {
     if (_voltage == 0) getVoltage();
-    if (_voltage > 0) _voltage_multiplier *= ((double) value / _voltage);
+    if (_voltage > 0) _voltage_multiplier *= (value / _voltage);
 }
 
-void HLW8012::expectedActivePower(unsigned int value) {
+void HLW8012::expectedActivePower(double value) {
     if (_power == 0) getActivePower();
-    if (_power > 0) _power_multiplier *= ((double) value / _power);
+    if (_power > 0) _power_multiplier *= (value / _power);
 }
 
 void HLW8012::resetMultipliers() {
